third_party/WebKit/LayoutTests/fast/dom/shadow/ancestor-pseudo-class-expected.txt

Test whether :ancestor matches a shadow host correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


:ancestor out of shadow tree should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:ancestor with * should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
:ancestor with tag selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:ancestor with class selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:ancestor with id selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:ancestor with attribute selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:ancestor in a shadow tree should match its shadow host.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
:ancestor with :host in a shadow tree should match its shadow host.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
:ancestor takes simple selectors and matches when one of the simple selectors matches.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
:ancestor matches a shadow host in just a nested shadow tree, not all enclosing shadow trees.
PASS backgroundColorOf('host1') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host1/host2') is "rgb(0, 128, 0)"
:ancestor matches based on a composed tree.
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
:ancestor matches based on a composed tree when having multiple shadow roots.
PASS backgroundColorOf('host/targetA') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host/targetB') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host/targetC') is "rgb(0, 128, 0)"
:ancestor is updated when its matched ancestor changes className or id.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
Compare :ancestor with :ancestor.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host') is "rgb(255, 0, 0)"
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
